Make Gradle metadata parsing lenient to errors
This commit makes metadata parsing lenient: if a version is not
the expected one, but we still manage to parse something, then
we don't fail. However, if parsing fails, we warn that the version
found in the metadata file is different from the one we expected.

This should give older Gradle versions a chance to parse newer
versions of Gradle metadata, especially in case new things are
added, that current Gradle doesn't care about.